Update model validations.

+++ db/migrate/20180113191852_create_users.rb
@@ -5,9 +5,9 @@ class CreateUsers < ActiveRecord::Migration[5.1]
       t.string :last_name
       t.string :email_address
       t.integer :age
-      t.decimal :height_feets
-      t.decimal :height_inches
-      t.decimal :weight
+      t.integer :height_feets
+      t.integer :height_inches
+      t.integer :weight
       t.string :favorite_color
 
       t.timestamps
